+++ apps/uiservice/src/components/sidebar/user-nav.tsx
@@ -1,14 +1,13 @@
 "use client"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Button } from "@/components/ui/button"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuGroup, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuShortcut, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
 import { GlobalContext } from "@/context/globalstate.provider"
-import { CircleUser, User } from "lucide-react"
+import { User } from "lucide-react"
 import { useRouter } from "next/navigation"
 import { useContext } from "react"
 
 export function UserNav() {
-  const [{ userState }] = useContext(GlobalContext)
+  const [{ user }] = useContext(GlobalContext)
   const router = useRouter()
 
   const signOut = async () => {
@@ -30,9 +29,9 @@ export function UserNav() {
       <DropdownMenuContent className="w-56" align="end" forceMount>
         <DropdownMenuLabel className="font-normal">
           <div className="flex flex-col space-y-1">
-            <p className="text-sm font-medium leading-none">{userState.name}</p>
+            <p className="text-sm font-medium leading-none">{user.name}</p>
             <p className="text-xs leading-none text-muted-foreground">
-              {userState.email}
+              {user.email}
             </p>
           </div>
         </DropdownMenuLabel>